WIP - style management and management rules
