+++ client/src/features/tags/components/TagsDisplay.jsx
@@ -1,12 +1,13 @@
 // Composant d'affichage pur
-
 import { useTags } from '../hooks/useTags';
 import { Plus, Pencil, Trash } from 'lucide-react';
-import { Badge } from "@/components/ui/badge"
+import { Badge } from "@/components/ui/badge";
 import { Button } from '@/components/ui/button';
-import { Input } from "@/components/ui/input"
+import { Input } from "@/components/ui/input";
 import { useState } from 'react';
-import { useCreateTag, useDeleteTag, useUpdateTag } from '../hooks/useMutateTag';
+import { useCreateTag } from '../hooks/useCreateTag';
+import { useUpdateTag } from '../hooks/useUpdateTag';
+import { useDeleteTag } from '../hooks/useDeleteTag';
 
 
 export default function TagsDisplay() {
@@ -18,6 +19,7 @@ export default function TagsDisplay() {
     const [tagName, setTagName] = useState("");
     const [editingTagId, setEditingTagId] = useState(null);
     const [editingName, setEditingName] = useState("");
+    const [isCreating, setIsCreating] = useState(false);
 
     if (isPending) {
         return <span> Loading ...</span>
@@ -41,7 +43,7 @@ export default function TagsDisplay() {
                 <h3>Tous mes tags :</h3>
                 {
                     data.map((tag) => (
-                        <div key={tag.id}>
+                        <div key={tag.id} className="flex items-center">
                             {editingTagId === tag.id ? (
                                 <>
                                     <Input value={editingName} onChange={ (e) => setEditingName(e.target.value)} />
@@ -54,7 +56,7 @@ export default function TagsDisplay() {
                                  <Badge style={{ backgroundColor: tag.color }}>{ tag.name }</Badge>
                                     {tag.by_default === false && (
                                         <>
-                                            <Button onClick={ () => {setEditingTagId(tag.id);setEditingName(tag.name)}}>< Pencil/></Button>
+                                            <Button onClick={ () => {setEditingTagId(tag.id);setEditingName(tag.name)}}>< Pencil size="icon" /></Button>
                                             <Button onClick={ () => deleteMutate(tag.id)}> <Trash /></Button> 
                                         </>
                                     ) 
@@ -66,8 +68,19 @@ export default function TagsDisplay() {
                         </div>
                     ))
                 }
-                <Input value={tagName} onChange={ (e) => setTagName(e.target.value) }/>
-                <Button onClick={ () => createMutate({ name: tagName })}>Ajouter un tag<Plus/></Button>
+                { isCreating === false ? (
+                    <Button onClick={ (e) => setIsCreating(true) }>Ajouter un tag<Plus/></Button>
+                ) : (
+                    <>
+                        <Input placeholder="Nom du tag"  className="w-48" value={tagName} onChange={ (e) => setTagName(e.target.value) }/>
+                            { createIsError  && (
+                                <p className="text-red-500"> {createError.message} </p>
+                            )}
+                        <Button onClick={ () => { createMutate({ name: tagName.trim() }, {onSuccess: () => {setTagName(""); setIsCreating(false)}}) }} disabled={tagName.trim() === "" | tagName.length < 2} > Valider </Button>
+                        <Button onClick={ () => { setIsCreating(false); setTagName("") }}>Annuler</Button>
+                    </>
+                )}
+
             </div>
         </>
     );
