--- client/src/features/tags/hooks/useMutateTag.js
@@ -1,65 +1,0 @@
-import { useQueryClient, useMutation } from '@tanstack/react-query';
-import { addNewTag, deleteTag, updateTag } from '../../../shared/services/tags_api';
-
-
-export const useCreateTag = () => {
-    const queryClient = useQueryClient();
-
-    const { mutate, isPending, isError, error } = useMutation({
-        mutationFn: addNewTag,
-        onSuccess: () => queryClient.invalidateQueries(
-            {
-                queryKey: ['tags']
-            }
-        )   
-    })
-
-    return {
-        mutate,
-        isPending,
-        isError,
-        error
-    }
-};
-
-
-export const useUpdateTag = () => {
-    const queryClient = useQueryClient();
-
-    const { mutate, isPending, isError, error } = useMutation({
-        mutationFn: ({ tagId, tagData }) => updateTag(tagId, tagData),
-        onSuccess: () => queryClient.invalidateQueries(
-            {
-                queryKey: ['tags']
-            }
-        )   
-    })
-
-    return {
-        mutate,
-        isPending,
-        isError,
-        error
-    }
-}
-
-
-export const useDeleteTag = () => {
-    const queryClient = useQueryClient();
-
-    const { mutate, isPending, isError, error } = useMutation({
-        mutationFn: deleteTag,
-        onSuccess: () => queryClient.invalidateQueries(
-            {
-                queryKey: ['tags']
-            }
-        )   
-    })
-
-    return {
-        mutate,
-        isPending,
-        isError,
-        error
-    }
-}
